# docs/openapi/book-openapi.yml
openapi: 3.0.3

info:
  title: Book API
  description: API for book management
  version: 1.0.0

servers:
  - url: http://localhost:8080/api
    description: Local server

tags:
  - name: Books
    description: Operations on books

paths:
  /books:
    get:
      tags:
        - Books
      summary: Retrieve all books
      operationId: getAllBooks
      parameters:
        - name: page
          in: query
          description: Page number (starting from 0)
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0

        - name: size
          in: query
          description: Number of items per page
          required: false
          schema:
            type: integer
            minimum: 1
            default: 20

        - name: sort
          in: query
          description: >
            Sorting criteria in the format `field,(asc|desc)`.
            Can be specified multiple times.
            Examples: `title,asc`, `author,desc`
          required: false
          explode: true
          schema:
            type: array
            items:
              type: string
          example:
            - title,asc
            - author,desc
      responses:
        '200':
          description: List of books retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookPage'

    post:
      tags:
        - Books
      summary: Create a new book
      operationId: createBook
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBookRequest'
      responses:
        '201':
          description: Book created successfully
          headers:
            Location:
              schema:
                type: string
        '409':
          description: A book with the same ISBN already exists

  /books/{isbn}:
    parameters:
      - name: isbn
        in: path
        required: true
        description: ISBN code of the book
        schema:
          type: string
          example: "9780132350884"

    get:
      tags:
        - Books
      summary: Retrieve a book by ISBN
      operationId: getBookByIsbn
      responses:
        '200':
          description: Book found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookResponse'
        '404':
          description: Book not found

    put:
      tags:
        - Books
      summary: Update an existing book
      operationId: updateBook
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateBookRequest'
      responses:
        '200':
          description: Book updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookResponse'
        '404':
          description: Book not found

    delete:
      tags:
        - Books
      summary: Delete a book
      operationId: deleteBook
      responses:
        '204':
          description: Book deleted successfully
        '404':
          description: Book not found

components:

  schemas:
    BookDto:
      type: object
      required:
        - bookType
        - isbn
        - title
        - author
      properties:
        bookType:
          type: string
          description: Determines the specific type of book
        isbn:
          type: string
          example: "9780132350884"
        title:
          type: string
          example: Clean Code
        author:
          type: string
          example: Robert C. Martin
        price:
          type: number
      discriminator:
        propertyName: bookType
        mapping:
          PAPERBACK: '#/components/schemas/PaperbackDto'
          EBOOK: '#/components/schemas/EBookDto'
          AUDIOBOOK: '#/components/schemas/AudioBookDto'

    PaperbackDto:
      allOf:
        - $ref: '#/components/schemas/BookDto'
        - type: object
          properties:
            pages:
              type: integer
              description: Number of physical pages
              example: 350
            weightGrams:
              type: number
              description: Weight of the book in grams
              example: 450.5

    EBookDto:
      allOf:
        - $ref: '#/components/schemas/BookDto'
        - type: object
          properties:
            format:
              type: string
              example: "EPUB"
            fileSizeMb:
              type: number
              example: 5.4

    AudioBookDto:
      allOf:
        - $ref: '#/components/schemas/BookDto'
        - type: object
          properties:
            narrator:
              type: string
              example: "Morgan Freeman"
            durationMinutes:
              type: integer
              example: 360

    BookResponse:
      oneOf:
        - $ref: '#/components/schemas/EBookDto'
        - $ref: '#/components/schemas/AudioBookDto'
        - $ref: '#/components/schemas/PaperbackDto'

    CreateBookRequest:
      oneOf:
        - $ref: '#/components/schemas/EBookDto'
        - $ref: '#/components/schemas/AudioBookDto'
        - $ref: '#/components/schemas/PaperbackDto'

    UpdateBookRequest:
      type: object
      properties:
        title:
          type: string
        author:
          type: string
        publicationDate:
          type: string
          format: date

    BookPage:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/BookResponse'

        pageable:
          $ref: '#/components/schemas/Pageable'

        totalElements:
          type: integer
          format: int64
          example: 100

        totalPages:
          type: integer
          example: 5

        size:
          type: integer
          example: 20

        number:
          type: integer
          example: 0

        numberOfElements:
          type: integer
          example: 20

        first:
          type: boolean
          example: true

        last:
          type: boolean
          example: false

        empty:
          type: boolean
          example: false

    Pageable:
      type: object
      properties:
        pageNumber:
          type: integer
          example: 0

        pageSize:
          type: integer
          example: 20

        offset:
          type: integer
          format: int64
          example: 0

        paged:
          type: boolean
          example: true

        unpaged:
          type: boolean
          example: false